+++ src/main/java/com/example/chackchack/common/security/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.example.chackchack.config;
+package com.example.chackchack.common.security;
 
 import com.example.chackchack.domain.common.dto.AuthUser;
 import com.example.chackchack.domain.user.enums.UserRole;
