+++ src/main/java/com/example/chackchack/common/security/JwtAuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.example.chackchack.config;
+package com.example.chackchack.common.security;
 
 import com.example.chackchack.domain.common.dto.AuthUser;
 import org.springframework.security.authentication.AbstractAuthenticationToken;
