+++ src/main/java/com/example/chackchack/common/security/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.example.chackchack.config;
+package com.example.chackchack.common.security;
 
 import com.example.chackchack.domain.common.exception.ServerErrorCode;
 import com.example.chackchack.domain.common.exception.ServerException;
